+++ biogpt-util.h
@@ -20,6 +20,8 @@
 
 using word_pair = std::pair<std::string, std::string>;
 
+static const size_t MB = 4*1024*1024;
+
 template<typename T>
 static void read_safe(std::ifstream& infile, T& dest) {
     infile.read((char*)& dest, sizeof(T));
